pyproject.toml
[project]
name = "ai-read-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.136.0",
    "openai>=2.32.0",
    "paddleocr>=3.5.0",
    "paddlepaddle>=3.3.1",
    "paddlex[ocr]>=3.5.1",
    "pillow>=11.0.0",
    "pydantic-settings>=2.14.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "torch[cpu]>=2.11.0",
    "torchvision>=0.20.0",
    "transformers>=5.5.4",
    "uvicorn>=0.45.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
